requirements.txt
openai-whisper
ffmpeg-python
sounddevice
scipy
pandas
python-dotenv
elevenlabs
openai
together